defs: rewrite const_cast() to do type checking
Actually this introduces const_cast1(), const_cast2() and
const_cast3(). I do not see a way to write const_cast() so that it
automatically picks the right dereference level without causing a
compiler error. But that's ok, type-checking const_castN() is good.
const_cast() is kept for backwards compilation compatibility.