tests: add and test a split mul variant of HX_timespec_mul
HX_timespec_mul_SPL, too, was try-out code. On Intel 1st-gen Atom, it
can be up to 148% slower than the normal variant (when not
optimizing). Otherwise it is basically on par with the normal variant,
being somewhere around 3–5% slower.

Only in i586+-O2 does SPL beat the normal variant, also with 3–5%,
all of which is very close to the margin of measurement error (3%).