string: implement quoting rule table
A new quote type that does URI encoding per RFC 3986 is desired.
However, it is not sufficient to use HXQUOTE_REJECT-type behavior on
the Reserved Characters in all cases, as the meaning of characters
that are neither in the Reserved set nor Unreserved set is URI- and
API-specific.

	# -*- perl -*-
	LWP::UserAgent->new()->request(
		HTTP::Request->new(GET => "http://foo.de/?bar baz"));

Perl's LWP for example supports encoding leftover characters, while
CURL does not. In the latter case, bluntly encoding everything via
HXQUOTE_ACCEPT-type behavior on the Unreserved set seems easiest.